Adding fixtures to conftest
