+++ tests/conftest.py
@@ -1,12 +1,15 @@
 import os
+from pathlib import Path
 
 import pytest
 import requests
 from mvg import MVG
-
+import uuid
+import json
 
 import argparse
 import sys
+from requests import HTTPError
 
 # This version of conftest.py adds some really ugly code to
 # run the tests as integration tests by running like
@@ -36,6 +39,12 @@ sys.argv[1:] = notknownargs
 # Retrieve API version to test against
 version_session = MVG("https://api.beta.multiviz.com", "NO TOKEN")
 VIBIUM_VERSION = "v" + str(version_session.tested_api_version)
+
+VALID_TOKEN = os.environ["TEST_TOKEN"]
+
+# Test data and session setup
+REF_DB_PATH = Path.cwd() / "tests" / "test_data" / "mini_charlie"
+SOURCE_ID_WAVEFORM = uuid.uuid1().hex  # generate a unique source per testrun
 
 
 def is_responsive(url):
@@ -81,3 +90,38 @@ else:
     def vibium():
         """Ensure that HTTP service is up and responsive."""
         return args.host
+
+
+@pytest.fixture(scope="session")
+def session(vibium):
+
+    url = vibium
+    print("Overriding vibium function with url %s", url)
+    session = MVG(url, VALID_TOKEN)
+    # To make sure we start from a clean slate
+    # we delete our resource in case it exists
+    # All information including measurements
+    # will be removed
+    # TO DO delete all so, currently we only
+    # handle resource SOURCE_ID
+    try:
+        session.get_source(SOURCE_ID_WAVEFORM)
+        print(f"Deleting {SOURCE_ID_WAVEFORM}")
+        session.delete_source(SOURCE_ID_WAVEFORM)
+    except HTTPError:
+        print(f"Source {SOURCE_ID_WAVEFORM} does not exist")
+
+    return session
+
+
+@pytest.fixture()
+def waveform_source(session):
+    try:
+        m_file_name = REF_DB_PATH / "u0001" / "meta.json"
+        with open(m_file_name, "r") as json_file:
+            meta = json.load(json_file)
+        # create_source happy case
+        session.create_source(SOURCE_ID_WAVEFORM, meta)
+        yield SOURCE_ID_WAVEFORM
+    finally:
+        session.delete_source(SOURCE_ID_WAVEFORM)
